bsod: Dump a plain text log file instead of XML
Since the log file is to be read and interpreted by humans, and since
XML is hard to read for both humans and machines, just output the
crash log in a plain text file which is easy to parse for both.

It also gets rid of a lot of string mangling and XML voodoo, reducing
the amount of code and memory allocations while already crashing.
It won't make another two in-memory copies of the whole log buffer, for
example.

Remove the attempt to blame others using maintainer.info because this
never really worked anyway.

* Get rid of logOutput signal completely

Move the error message ringbuffer implementation to eerror.cpp from
bsod.cpp, and just expose the method to fetch the data. No need for
the extra libsig overhead if there's only one client.

Also move the (mandatory!) locking into the logOutput method, it locks
access to the ringbuffer.

* bsod: provide dmesg in crashlog

Provide the dmesg in enigma2 crashlog because it really helps to identify kernel crashes or other issues usually visible on dmesg.

* eDebug: simplify ringbuffer/logOutput calls

The enigma ringbufer does not use the loglevel information passed
to it. So remove it and get rid of overloaded addToLogbuffer function.
Reduce logOutput signal to use char * instead of std::string, to
avoid expensive std::string usage.
Can savely do this as eDebug is the only user of the ringbuffer.